ARS and SRS for all stocks in stocks.csv
Calculate the ARS and SRS values. The concept of ARS is based on selecting a threshold date and compare the relative progress of the instrument with respect to benchmark. Whereas the SRS is based on a fixed length, (123 - 6 months roughly). A stock that is positive ARS and SRS indicates that it is showing strength with respect to the strongest stocks in the entire market.