Green Dot Scan for Mean Reversion Expansion
This is a scanner for reversion expansion of a stock from its means. A "Green Dot" indicates expansion from its mean, which will help in long side trades if the stock is just moving out from its cluster of means. This script scans twice, once on daily timeframe and then on weekly. An expansion from weekly is generally more powerful and rewarding than from a daily. This scan, can be visually checked with the same indicator on TradingView. The indicator is listed here and open sourced in pinescript. https://in.tradingview.com/script/fUL2vFoP-Price-Mean-Reversion-Expansion/