supply exhaustion scan
This scan scans for stocks that are witnessing a supply exhaustion. The way I do it, is by finding a stock that hit its lowest close at least 6 months ago. Then we go back in time from that low date and find the highest price in a reasonable timeframe. We then take a fibonacci retracement upwards from the difference of this height and look for retracements below 50%, 38.2% and 23.6%. Sometimes it is good to catch stocks early in their movment, when they are turning up from a supply exhaustion. An increasing RS, along with confirmation of price and fundamental data, is a very probable set up once can trade. We can make good trades, by not trading just Highs and ATH too!