Multi Month BO
This scans closes that are near record high close of the stock. So technically we are not looking at All Time "High" but All Time High Close. Further, we look for such Closes that should have happened at least 11-12 months ago. We track down such stocks that are today trading between 5% of such High Closes. The trading strategies around this should be carefully built, because there is always a high probability of the breakout failing. Yet when there is a breakout and sustained it rewards handsomely.